Fix testing-common tests
