+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/matcher/AdditionalLibraryIgnoresMatcher.java
@@ -42,12 +42,15 @@ public class AdditionalLibraryIgnoresMatcher<T extends TypeDescription>
         || name.startsWith("com.jayway.jsonpath.")
         || name.startsWith("com.lightbend.lagom.")
         || name.startsWith("javax.el.")
-        || name.startsWith("net.sf.cglib.")
         || name.startsWith("org.apache.lucene.")
         || name.startsWith("org.apache.tartarus.")
         || name.startsWith("org.json.simple.")
         || name.startsWith("org.yaml.snakeyaml.")) {
       return true;
+    }
+
+    if (name.startsWith("net.sf.cglib.")) {
+      return !name.equals("net.sf.cglib.core.internal.LoadingCache$2");
     }
 
     if (name.startsWith("org.springframework.")) {
