+++ testing-common/integration-tests/src/test/groovy/InstrumentOldBytecode.groovy
@@ -5,10 +5,7 @@
 
 import com.ibm.as400.resource.ResourceLevel
 import io.opentelemetry.instrumentation.test.AgentTestRunner
-import spock.lang.Ignore
 
-// FIXME (trask)
-@Ignore
 class InstrumentOldBytecode extends AgentTestRunner {
   def "can instrument old bytecode"() {
     expect:
