+++ testing-common/integration-tests/src/test/groovy/server/ServerTest.groovy
@@ -11,11 +11,8 @@ import io.opentelemetry.instrumentation.test.AgentTestRunner
 import io.opentelemetry.instrumentation.test.utils.OkHttpUtils
 import okhttp3.MultipartBody
 import okhttp3.Request
-import spock.lang.Ignore
 import spock.lang.Shared
 
-// FIXME (trask)
-@Ignore
 class ServerTest extends AgentTestRunner {
   @Shared
   def client = OkHttpUtils.client()
