+++ testing-common/integration-tests/src/test/java/config/exclude/SomeClass.java
