+++ testing-common/integration-tests/src/test/java/config/exclude/packagename/SomeClass.java
