+++ testing-common/library-for-integration-tests/library-for-integration-tests.gradle
@@ -1,0 +1,1 @@
+apply plugin: 'java-library'
