+++ testing-common/library-for-integration-tests/src/main/java/library/DisabledKeyClass.java
@@ -1,0 +1,9 @@
+package library;
+
+/** A class that is used that field injection can be disabled. */
+public class DisabledKeyClass extends KeyClass {
+  @Override
+  public boolean isInstrumented() {
+    return false;
+  }
+}
