+++ testing-common/library-for-integration-tests/src/main/java/library/KeyClass.java
@@ -1,0 +1,31 @@
+package library;
+
+public class KeyClass {
+  public boolean isInstrumented() {
+    // implementation replaced with test instrumentation
+    return false;
+  }
+
+  public int incrementContextCount() {
+    // implementation replaced with test instrumentation
+    return -1;
+  }
+
+  public int incrementContextCountWithFactory() {
+    // implementation replaced with test instrumentation
+    return -1;
+  }
+
+  public int getContextCount() {
+    // implementation replaced with test instrumentation
+    return -1;
+  }
+
+  public void putContextCount(int value) {
+    // implementation replaced with test instrumentation
+  }
+
+  public void removeContextCount() {
+    // implementation replaced with test instrumentation
+  }
+}
