+++ testing-common/library-for-integration-tests/src/main/java/library/UntransformableKeyClass.java
@@ -1,0 +1,12 @@
+package library;
+
+/**
+ * A class which will not be transformed by our instrumentation due to {@link
+ * FieldBackedProviderTest#skipTransformationConditions()}.
+ */
+public class UntransformableKeyClass extends KeyClass {
+  @Override
+  public boolean isInstrumented() {
+    return false;
+  }
+}
